snapshot before testing

+++ app/Http/Controllers/ProjectController.php
@@ -2,16 +2,19 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Project;
 use App\Models\Task;
+use App\Models\Project;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
 
 class ProjectController extends Controller
 {
+    // STILL TO ADD USER AUTHORIZATION IN DELETE AND UPDATE
+
     //show all project , returns index 
     public function index(){
         return view('project.projects',[
-            //'projects' => Project::all()
+            'projects' => Project::all()
         ]);
     }
 
@@ -22,11 +25,12 @@ class ProjectController extends Controller
 
     //store project forum
     public function store(Request $request){
-        //TO FILL 
         $fields_to_store = $request->validate([
-            'pname'=>'required',
+            'pname'=>['required',Rule::unique('projects', 'pname')],
             'description'=>'nullable' //verify that nullable works
         ]); 
+
+        $fields_to_store['user_id'] = auth()->id();
 
         Project::create($fields_to_store);
         return redirect('project.projects');
@@ -40,6 +44,12 @@ class ProjectController extends Controller
 
     //update project request
     public function update(Request $request , Project $project){
+
+         //USER AUTHORIZATION
+         if($project->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $fields_to_store = $request->validate([
             'pname'=>'required',
             'description'=>'nullable' //verify that nullable works
@@ -51,6 +61,12 @@ class ProjectController extends Controller
 
     //delete project request
     public function delete(Project $project){
+
+         //USER AUTHORIZATION
+         if($project->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+        
         $project->delete();
         return redirect('project.projects');
     }
@@ -58,10 +74,11 @@ class ProjectController extends Controller
 
     //open a single project  , returns show
     public function show(Project $project){
+        $tasks = Project::find($project->id)->tasks; //still need to define relashionship
+
         return view('project.tasks',[
-            //'project' => $project
-            //'tasks' => Task::all()->where('') //STILL NOT DONE
-            //get tasks of project in $project
+            'project' => $project,
+            'tasks' => $tasks 
         ]);
     }
